Refactor Dashboard component to add NavbarDashboard, AsideDashboard, and DashboardLayout components
