+++ client/src/App.jsx
@@ -7,6 +7,8 @@ import DHAdmin from "@pages/DHAdmin.jsx";
 import Register from "@pages/Register.jsx";
 import Login from "@pages/Login.jsx";
 import Dashboard from "@pages/Dashboard.jsx";
+import PostDashboard from "@components/PostDashboard.jsx";
+import UserDashboard from "@components/UserDashboard.jsx";
 
 function App() {
     return (
@@ -20,7 +22,10 @@ function App() {
                 <Route path="/dh-admin" element={<DHAdmin />} />
                 <Route path="/dh-admin/register" element={<Register />} />
                 <Route path="/dh-admin/login" element={<Login />} />
+                
                 <Route path="/dh-admin/dashboard" element={<Dashboard />} />
+                <Route path="/dh-admin/dashboard/usersDashboard" element={<UserDashboard />} />
+                <Route path="/dh-admin/dashboard/postsDashboard" element={<PostDashboard />} />
             </Routes>
         </>
     );
