+++ client/src/components/DashboardStructure/NavbarDashboard.jsx
@@ -1,0 +1,6 @@
+const NavbarDashboard = () => {
+  return (
+    <div>NavbarDashboard</div>
+  )
+}
+export default NavbarDashboard
